requirements.txt
# Core Requirements
pyspark==3.5.0 
delta-spark==3.1.0

# Testing & Development
requests==2.31.0
tenacity==8.2.2
pytest==7.4.0
pytest-spark==0.6.0
pylint==2.17.4

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
pytz
#pandas==2.0.3  # For local testing

# Cloud Storage (Optional - include based on your deployment target)
# boto3==1.28.57  # For S3
# adlfs==2023.6.0  # For Azure Data Lake
# gcsfs==2023.6.0  # For Google Cloud Storage
